Introduce function, bpeo_is_ics().
This function is used to determine if we're currently on an ICS URL.

The previous logic sometimes introduced conflicts between an event
page whose slug is comprised solely of an hexadecimal digit and with a
check for the hash used for private iCalendar subscription URLs (those
whose URLs ended in /XXX/ical/).

Fixes #70.