iCal Sync: Register EO's add feed AJAX action with BuddyPress.
This is needed because BuddyPress 12 refactored their URI router. So
in order to use BuddyPress page conditionals on AJAX hooks, plugins
now need to register which AJAX actions will need access to BP URI
globals.

We need access to see if a BuddyPress group is adding the EO iCal
feed.